bot/requirements.txt
pyTelegramBotAPI
requests
PyMuPDF
Pillow
pytesseract
python-dotenv
psycopg2-binary